requirements.txt
pygetwindow
pyautogui
Pillow
requests
PyQt6
pyinstaller
python-dotenv